+++ package.json
@@ -1,51 +1,26 @@
 {
   "name": "canopy-webpack-config",
-  "type": "commonjs",
   "version": "4.3.1",
   "description": "Some defaults for webpack configs at canopy",
   "main": "src/canopy-webpack-config.js",
-  "repository": "git@github.com:CanopyTax/canopy-webpack-config.git",
-  "author": "Joel Denning <joel.denning@canopytax.com>",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@babel/core": "^7.28.0",
-    "@babel/preset-env": "^7.28.0",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
-    "@rollup/stream": "^3.0.1",
-    "@web/dev-server": "^0.4.6",
-    "@web/dev-server-hmr": "^0.4.1",
-    "chalk": "^5.4.1",
-    "grunt": "^1.6.1",
-    "grunt-cli": "^1.5.0",
-    "gulp": "^5.0.1",
-    "gulp-sourcemaps": "^3.0.0",
-    "postcss": "^8.5.6",
-    "prettier": "^3.6.2",
-    "vinyl-buffer": "^1.0.1",
-    "vinyl-source-stream": "^2.0.0",
-    "yargs": "^18.0.0"
+  "exports": {
+    ".": "./src/canopy-webpack-config.js",
+    "./esm": "./src/canopy-webpack-config-esm.mjs"
   },
+  "repository": "git@github.com:CanopyTax/canopy-webpack-config.git",
+  "license": "Apache-2.0",
   "peerDependencies": {
     "@babel/plugin-transform-runtime": ">=7",
     "@babel/runtime": ">=7"
   },
+  "devDependencies": {
+    "prettier": "^3.6.2"
+  },
   "dependencies": {
     "clean-webpack-plugin": "^4.0.0",
-    "url": "^0.11.4",
-    "webpack-bundle-analyzer": "^4.10.2",
-    "webpack-merge": "^6.0.1"
+    "url": "^0.11.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-merge": "^4.2.1"
   },
-  "optionalDependencies": {
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^28.0.6",
-    "@rollup/plugin-node-resolve": "^16.0.1",
-    "@rollup/plugin-replace": "^6.0.2",
-    "@rollup/plugin-terser": "^0.4.4",
-    "rollup": "^4.45.1",
-    "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-serve": "^3.0.0"
-  },
-  "packageManager": "yarn@4.9.2"
+  "packageManager": "yarn@3.6.2"
 }
